requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiohttp==3.9.5
pycryptodome==3.20.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.9
qrcode[pil]==7.4.2
Pillow==10.3.0
loguru==0.7.2
